+++ helpers/dataManipulation.js
@@ -119,6 +119,72 @@ export const castHeadingLevelFromMixed = (val: { [string]: ?mixed }, key: string
 }
 
 /**
+ * Simple Object equality test, working for ONE-LEVEL only objects.
+ * from https://stackoverflow.com/a/5859028/3238281
+ * @param {Object} o1
+ * @param {Object} o2
+ * @returns {boolean} does o1 = o2?
+ * @test in jest file
+ */
+export function compareObjects(o1: Object, o2: Object): boolean {
+  for (const p in o1) {
+    if (o1.hasOwnProperty(p)) {
+      if (o1[p] !== o2[p]) {
+        return false
+      }
+    }
+  }
+  for (const p in o2) {
+    if (o2.hasOwnProperty(p)) {
+      if (o1[p] !== o2[p]) {
+        return false
+      }
+    }
+  }
+  return true
+}
+/**
+ * Remove the 'exclude' array terms from given 'arr' array.
+ * Assumes both arrays are of the same Object type, and that we will only remove
+ * when all properties are equal.
+ * @param {Array<Object>} arr - array to remove from
+ * @param {Array<Object>} exclucde - array to remove
+ * @returns {Array<Object>} arr minus exclude
+ * @tests in jest file
+ */
+
+export function differenceByObjectEquality<P: string, T: { +[P]: mixed }> (
+  arr: $ReadOnlyArray < T >,
+    exclude: $ReadOnlyArray < T >
+): Array < T > {
+  return arr.filter(
+    (a: T) => !exclude.find((b: T) => compareObjects(b, a))
+  )
+}
+
+/**
+ * Compute difference of two arrays, by a given property value
+ * from https://stackoverflow.com/a/63745126/3238281
+ * translated into Flow syntax with Generics by @nmn:
+ * - PropertyName is no longer just a string type. It's now a Generic type itself called P. But we constrain P such that it must be string. How is this different from just a string? Instead of being any string, P can be a specific string literal. eg. id
+ * - T is also constrained. T can no longer be any arbitrary type. It must be an object type that contains a key of the type P that we just defined. It may still have other keys indicated by the ...
+ * @param {<Array<T>} arr The initial array
+ * @param {<Array<T>} exclude The array to remove
+ * @param {string} propertyName the key of the object to match on
+ * @return {Array<T>}
+ * @tests in jest file
+ */
+export function differenceByPropVal<P: string, T: { +[P]: mixed, ... }>
+  (arr: $ReadOnlyArray < T >,
+    exclude: $ReadOnlyArray < T >,
+      propertyName: P
+  ): Array < T > {
+  return arr.filter(
+    (a: T) => !exclude.find((b: T) => b[propertyName] === a[propertyName])
+  )
+}
+
+/**
  * Recursively rename keys at any level of an object
  * Written by cursor.ai
  * @param {any} obj